refactor(type): split combat signal payload types
