+++ Source/Portfolio/Type/CWeaponStructure.h
@@ -8,6 +8,8 @@
 #include "Type/CCombatHitTypes.h"
 #include "Type/CCombatDamageTypes.h"
 #include "Type/CCombatResultTypes.h"
+#include "Type/CCombatSignalSourceTypes.h"
+#include "Type/CCombatSignalTargetTypes.h"
 #include "Type/CStateTypes.h"
 #include "Type/CHealthTypes.h"
 #include "CWeaponStructure.generated.h"
@@ -233,44 +235,6 @@ enum class EActionRequestRejectReason : uint8
 };
 
 UENUM(BlueprintType)
-enum class ECombatSignalSourceRejectReason : uint8
-{
-	None = 0,
-
-	InvalidRequest,
-
-	InvalidAttacker,
-	InvalidDamageCauser,
-	InvalidTarget,
-	InvalidInstigator,
-
-	SpecNotFound,
-	ComputeFailed,
-	CommitFailed,
-
-	// Reject Reason of 'CanSendCombatSignal'
-	InvalidOwner,
-	SelfTarget,
-	DuplicateHitInWindow,
-	FriendlyTarget,
-};
-
-UENUM(BlueprintType)
-enum class ECombatSignalTargetRejectReason : uint8
-{
-	None = 0,
-
-	InvalidTarget,
-	InvalidCauser,
-	InvalidInstigator,
-
-	AlreadyDead,
-	ZeroDamage,
-
-	UnknownCueTag,
-};
-
-UENUM(BlueprintType)
 enum class EReactionStopSource : uint8
 {
 	None = 0,
@@ -451,276 +415,6 @@ public:
 
 public:
 	bool IsValidMinimal() const;
-};
-
-USTRUCT(BlueprintType)
-struct FCombatSignalSourcePayload
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	FHitContext HitContext = FHitContext();
-
-	UPROPERTY(Transient)
-	AActor* SourceActor = nullptr;
-
-	UPROPERTY(Transient)
-	AActor* DamageCauser = nullptr;
-
-	UPROPERTY(Transient)
-	AActor* TargetActor = nullptr;
-
-	UPROPERTY(Transient)
-	FCombatSignalHitWindowKey HitWindowKey = FCombatSignalHitWindowKey();
-
-	UPROPERTY(Transient)
-	FDamageImpactInfo DamageImpactInfo = FDamageImpactInfo();
-
-	UPROPERTY(Transient)
-	FDamageSpecKey DamageSpecKey = FDamageSpecKey();
-
-public:
-	FCombatSignalSourcePayload() = default;
-};
-
-USTRUCT(BlueprintType)
-struct FCombatSignalSourceContext
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	bool bAccepted = true;
-
-	UPROPERTY(Transient)
-	ECombatSignalSourceRejectReason RejectReason = ECombatSignalSourceRejectReason::None;
-
-	UPROPERTY(Transient)
-	FHitContext HitContext = FHitContext();
-
-	UPROPERTY(Transient)
-	AActor* SourceActor = nullptr;
-
-	UPROPERTY(Transient)
-	AController* Instigator = nullptr;
-
-	UPROPERTY(Transient)
-	AActor* DamageCauser = nullptr;
-
-	UPROPERTY(Transient)
-	AActor* TargetActor = nullptr;
-
-	UPROPERTY(Transient)
-	FCombatSignalHitWindowKey HitWindowKey = FCombatSignalHitWindowKey();
-
-	UPROPERTY(Transient)
-	FDamageImpactInfo DamageImpactInfo = FDamageImpactInfo();
-
-	UPROPERTY(Transient)
-	FDamageSpecKey DamageSpecKey = FDamageSpecKey();
-
-	UPROPERTY(Transient)
-	FDamageSpec DamageSpec = FDamageSpec();
-
-	UPROPERTY(Transient)
-	FDamageAmount DamageAmount = FDamageAmount();
-
-	UPROPERTY(Transient)
-	float CommittedDamage = 0.f;
-
-public:
-	FCombatSignalSourceContext() = default;
-};
-
-USTRUCT(BlueprintType)
-struct FCombatSignalSourceResult
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	bool bAccepted = true;
-
-	UPROPERTY(Transient)
-	ECombatSignalSourceRejectReason RejectReason = ECombatSignalSourceRejectReason::None;
-
-	UPROPERTY(Transient)
-	FCombatSignalHitWindowKey HitWindowKey = FCombatSignalHitWindowKey();
-
-	UPROPERTY(Transient)
-	FDamageSpecKey DamageSpecKey = FDamageSpecKey();
-
-	UPROPERTY(Transient)
-	float BaseDamage = 0.f;
-
-	UPROPERTY(Transient)
-	float RequestDamage = 0.f;
-
-	UPROPERTY(Transient)
-	float CommittedDamage = 0.f;
-
-public:
-	FCombatSignalSourceResult() = default;
-};
-
-USTRUCT(BlueprintType)
-struct FCombatSignalTargetPayload
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	class AActor* SourceActor = nullptr;
-
-	UPROPERTY(Transient)
-	class AActor* TargetActor = nullptr;
-
-	UPROPERTY(Transient)
-	class AController* EventInstigator = nullptr;
-
-	UPROPERTY(Transient)
-	class AActor* DamageCauser = nullptr;
-
-	UPROPERTY(Transient)
-	FDamageImpactInfo DamageImpactInfo = FDamageImpactInfo();
-
-	UPROPERTY(Transient)
-	FDamageSpecKey DamageSpecKey = FDamageSpecKey();
-
-	UPROPERTY(Transient)
-	FDamageSpec DamageSpec = FDamageSpec();
-
-	UPROPERTY(Transient)
-	FDamageAmount DamageAmount = FDamageAmount();
-
-	UPROPERTY(Transient)
-	float RequestedDamage = 0.f;
-
-public:
-	FCombatSignalTargetPayload() = default;
-};
-
-USTRUCT(BlueprintType)
-struct FCombatSignalTargetContext
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	class AActor* SourceActor = nullptr;
-
-	UPROPERTY(Transient)
-	class AActor* TargetActor = nullptr;
-
-	UPROPERTY(Transient)
-	class AController* Instigator = nullptr;
-
-	UPROPERTY(Transient)
-	class AActor* DamageCauser = nullptr;
-
-	UPROPERTY(Transient)
-	FDamageImpactInfo DamageImpactInfo = FDamageImpactInfo();
-
-	UPROPERTY(Transient)
-	FDamageSpecKey DamageSpecKey = FDamageSpecKey();
-
-	UPROPERTY(Transient)
-	bool bAccepted = true;
-
-	UPROPERTY(Transient)
-	ECombatSignalTargetRejectReason RejectReason = ECombatSignalTargetRejectReason::None;
-
-	UPROPERTY(Transient)
-	EDamageDefenseOutcome DefenseOutcome = EDamageDefenseOutcome::None;
-
-	UPROPERTY(Transient)
-	bool bShouldCommitDamage = true;
-
-	UPROPERTY(Transient)
-	float HealthPointBefore = 0.f;
-
-	UPROPERTY(Transient)
-	EDeadState DeadState_Before = EDeadState::Alive;
-
-	UPROPERTY(Transient)
-	float RequestedDamage = 0.f;		// Raw incoming damage requested by Apply pipeline. (ex. [skill] 100)
-
-	UPROPERTY(Transient)
-	float MitigatedDamage = 0.f;		// Post-mitigation damage after target defenses. (ex. [guard/resistance] 100 -> 70)
-
-	UPROPERTY(Transient)
-	float FinalTakenDamage = 0.f;		// Final damage decided by Take evaluation rules. (ex. [clamp max/min damage-limit] 70 -> 60)
-
-	UPROPERTY(Transient)
-	float CommittedDamage = 0.f;		// Actual HP loss committed to Health. (ex. [shield absorbs] 60 -> HP: -30 / SP: -30)
-
-	UPROPERTY(Transient)
-	float HealthPointAfter = 0.f;
-
-	UPROPERTY(Transient)
-	EDeadState DeadState_After = EDeadState::Alive;
-
-public:
-	FCombatSignalTargetContext() = default;
-};
-
-USTRUCT(BlueprintType)
-struct FCombatSignalTargetResult
-{
-	GENERATED_BODY()
-
-public:
-	UPROPERTY(Transient)
-	bool bAccepted = true;
-
-	UPROPERTY(Transient)
-	ECombatSignalTargetRejectReason RejectReason = ECombatSignalTargetRejectReason::None;
-
-	UPROPERTY(Transient)
-	EDamageDefenseOutcome DefenseOutcome = EDamageDefenseOutcome::None;
-
-	UPROPERTY(Transient)
-	bool bShouldCommitDamage = true;
-
-	UPROPERTY(Transient)
-	FDamageSpecKey DamageSpecKey = FDamageSpecKey();
-
-	UPROPERTY(Transient)
-	float RequestDamage = 0.f;
-
-	UPROPERTY(Transient)
-	float MitigatedDamage = 0.f;
-
-	UPROPERTY(Transient)
-	float FinalTakenDamage = 0.f;
-
-	UPROPERTY(Transient)
-	float CommittedDamage = 0.f;
-
-	UPROPERTY(Transient)
-	EDeadState DeadState_Before = EDeadState::Alive;
-
-	UPROPERTY(Transient)
-	EDeadState DeadState_After = EDeadState::Alive;
-
-public:
-	FCombatSignalTargetResult() = default;
-};
-
-USTRUCT(BlueprintType)
-struct FCombatSignalTargetPacket
-{
-	GENERATED_BODY()
-
-	UPROPERTY(Transient)
-	FCombatSignalTargetPayload Payload;
-
-	UPROPERTY(Transient)
-	FCombatSignalTargetContext Context;
-
-	UPROPERTY(Transient)
-	FCombatSignalTargetResult Result;
 };
 
 USTRUCT(BlueprintType)
